Linear Regression intro
A basic linear regression model with test values. The data used is used to display a general idea of the principal. We created numpy arrays and utilized LinearRegression module from SKLEARN. This allows us to build our model simply. Once the model is created we fit these and then predict values  from x. Note: As this is a test, we use the same data as we used to fit. In a real world example this would NOT be advised. In subsequent testing we split the data into test and train.