Linear Regression enhanced 2
For this example we utilize sklearn's modules of mean_squared_error, r2_score and PolynomialFeatures.

Using these allow us to create multiple transformed objects with varying degree to square and transform our x values. We then use these values to fit to seperate models and compare the square root of the mean_squared_errors  of each model. Finally to visualzie this data we plot all the values using matplot.